Update Tags to have its own component
