+++ src/components/CharactersTable/CharactersTable.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react'
 import type { Character, CharacterAbility, CharacterTag } from '../../types'
+import Tags from '../Tags/Tags'
 import {
   useReactTable,
   getCoreRowModel,
@@ -15,24 +16,6 @@ const columnHelper = createColumnHelper<Character>()
 //   }
 //   return false
 // }
-
-const Tags = ({ tags }: { tags: CharacterTag[] }) => {
-  if (!tags) {
-    return <></>
-  }
-
-  return (
-    <>
-      {tags.map((tag) => {
-        return (
-          <span key={tag.slot} className='badge'>
-            {tag.tag_name}
-          </span>
-        )
-      })}
-    </>
-  )
-}
 
 const columns = [
   columnHelper.accessor('name', {
