+++ src/components/Tags/Tags.css
@@ -1,0 +1,13 @@
+.badge {
+  background-color: #ffffff;
+  color: #207a9b;
+  margin-right: 4px;
+  padding: 4px 8px;
+  border-radius: 12px;
+  border-style: solid;
+  border-color: #207a9b;
+  border-width: 1px;
+  font-size: 12px;
+  font-weight: bold;
+  text-transform: uppercase;
+}
